Recognize Home activity started from different uids
Home activity can be started from uid different from current uid and
zero. It may happen when new user is added or when Home activity is
killed by the system and then started from Recents activity.

In this case new ActivityRecord for Home activity will have type
APPLICATION_ACTIVITY_TYPE and then it will be placed to application
activity stack instead for Home activity stack.

If it happens and then Home button is pressed, one more Home activity
will be started and placed to Home stack. Home activity has
SingleInstance flag, and if more than one instance is present, it may
behave in a weird way. Usualy it crashes.

From user perspective, this issue is visible as a crash in Home
application.

Issue is fixed by removing all constraints for _launchedFromUid when
activity type is defined.

Change-Id: I260e4e06a678142de477967a2d5bbf4f35560ec1