Fixed inconsistent result in getPreferredActivities function
According to the documentation, PackageManager.getPreferredActivities()
function should return the list of all preferred activities for the
given package name. In case when the package name is null, preferred
activities for all packages should be returned.
The inconsistency occures when the user selects activity in
ResolverActivity to be used just once. At this point,
getPreferredActivities will return the selected activity if package
name is null and will not return it if package name is specified.

The problem is fixed by applying the same check regardless if package
name is specified or not.

Change-Id: Ia9ffc25ffc1f2cf59b932217a81bd09c0a159e82